[roll] Roll fuchsia [fint] Call gn gen from fint set
Without any args for now, to keep this CL small. I'll start adding args
in a follow-up.

The fuchsia-x64-release-fint builder is still not expected to pass with
this change, because we're not yet setting the necessary args to produce
all the artifacts that infra expects from gn.

Original-Bug: 61977
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/438962
Original-Revision: 79d38bd0583fa64b9d98ef44aae62f9e49579a75
GitOrigin-RevId: 17be14add4818d1a2422ab22205cc60258a4401a
Change-Id: Ia40f2cbd45df55408c5b89923020c12c77ebf4cc