[roll] Roll fuchsia [testrunner] Execute a SEV instruction during SSH diags
Trigger a kernel command which will cause a core to execute a SEV
instruction (send event) when running the diags attempting to figure
out why SSH is non-responsive.

Some of the recent diags results indicate that one of the causes of
SSH being non-responsive might be the issue we have with cores
apparently becoming stuck in WFE on Sherlock while attempting to
obtain a kernel spinlock.

If this is the case, then sending an event with SEV should unstick the
core, and we should immediately see its thread queue drain.

Original-Bug: 59720
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/437974
Original-Revision: 5ca996277910ee706138ff00a380fee71fdab32b
GitOrigin-RevId: f0e90a6de05bb1567e7641c963f567ce3ed4e33d
Change-Id: I8522c37ff699aa7c2763801f27d8a22128211114