[roll] Roll fuchsia [e2e] Add a timezone test for flutter
This is similar to fxr/367445, but using the flutter runner this time.
While the mechanims for obtaining the locales are identical to those in
the dart test, the runner is still a separate binary so could be subject
to independent variation. So we redo the testing.

This commit factors the common parts of the test into a separate rust
library.

Original-Bug: 43341
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/391701
Original-Revision: 92811f5da5c4c2bb48b64d15a3b56a9ade5ebc03
GitOrigin-RevId: b25e8f4e2fd697a2387cc4c7015adac57005f50e
Change-Id: I01bf6298cf8189c37d1559ceb87ed133ba8968ce