[roll] Roll fuchsia [settings] Add broker filter in policy proxy and implement direct responses
This CL filters the messages sent to policy proxies for only messages
that are sent by the switchboard and that are requests for a particular
setting type. These messages are forwarded to the policy handler to take
action.

The pass through case is trivially implemented by doing nothing with the
message, so that it continues onto the setting handler as normal. The
direct response case is also implemented, where the policy handler
directly gives a response to the client and the setting handler never
receives a message. The modified request case where the policy handler
modifies the request and passes it onto the setting handler is not yet
implemented.

Original-Bug: 59747
Test: fx test setui_service_tests
MULTIPLY: setui_service_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/438577
Original-Revision: 5ee0272b9bdf47bf31d2a5a04240c7729df60933
GitOrigin-RevId: b5dc02d84b9823ece55f8518635c26112929013e
Change-Id: I16a35ed107d238b8b6e2da5853081e1dc7abafdd