[roll] Roll fuchsia [fidlgen][cpp] Use FullDecl to disambiguate types
In C++, it is ambiguous to use the name `Foo` if there is both a method
and a class named `Foo` in scope. The existing approach in the codegen
to disambiguate them attempts to prefix the class ones with `class Foo`,
but won't work for all cases (e.g. it would treat `::std::unique_ptr<Foo>`
as already fully qualified when it isn't). An easier alternative is to
just always refer to types by their fully qualified name.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/438383
Original-Revision: ebfdd3e4086f5eec909080780059724a4b12901c
GitOrigin-RevId: 1434988628a26c25f52b586362372d7f2b114792
Change-Id: I52a7904851d0a399789d9dde51cf8f31a3caeab3