[roll] Roll infra/recipes [cipd_roller] Make failed cipd describe steps green
If a `cipd describe` step in a CIPD roller fails because the package did
not have the requested version, we're generally able to recover by
falling back to requesting the same version as one of the other
packages. So it's confusing to present it as a failed step in the Milo
UI. Instead, we should present it as green, and only show a red step at
the end of the build if we're unable to find a common tag to roll.

Original-Bug: 62236
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/440996
Original-Revision: 80b01c67f158aca7d74e4cbe6619999e3365f462
GitOrigin-RevId: d0134c65380da39223f589f6bf8b1c2a83c9bb7b
Change-Id: I18fbbd6cf02eba720d74e9d766766285b4ead05b