[roll] Roll fuchsia [perfmon][trace_converters] Don't replace LBR records with heap garbage
The calculation of the size of a LBR record uses + instead of * to
calculate an array size. Additionally, the chromium exporter is storing
LBR records in a vector, which can't work as they are variable size.

Also adds a few missing fields of each LBR record to the exported JSON.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/438517
Original-Revision: 1bf6d5e9a7f15419131a88c8b94530f92955a67c
GitOrigin-RevId: 03259250ae33b52e25b65b7b8f1d3ca6fd78cf07
Change-Id: I40be0e1754458da3147ebcd649b52ee47dde180a