[roll] Roll infra/recipes [build] Delete ToolchainContext
ToolchainContext added an unnecessary and confusing layer of indirection
to the GN arg construction logic and to the ninja setup.

Also, instead of storing the final `use_goma` boolean as a global
attribute of the ToolchainContext, have it be a parameter to the build
module functions that require it. This means that we can no longer
set the global `use_goma` value to false if the gcc or clang toolchain
is overridden, so we'll have to explicitly set `use_goma=False` for all
gcc and clang toolchain builders before landing this.

Original-Bug: 61977
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/441074
Original-Revision: 8f34eb1e92dde1045a1c6d0318b2b8745847717a
GitOrigin-RevId: aa76a663ad30826c7ab7e9c721c0e54c40f0dba7
Change-Id: I03cad9ca6451862d5694f63586e857ce85f53dae