[roll] Roll fuchsia [fidl][benchmarks] Setup benchmarks for fidlc
During the union to xunion migration, there were performance
regressions in FIDLC for certain inputs. This CL adds
benchmarks to help track and quantify the regressions.
The JSON generation path is the target.

In addition to tracking regressions, there are new usages
of fidlc such as language server support for IDEs that
depend on fast compilation performance and this opens
the door to optimizing for that case.

Only a single benchmark was added and chromeperf pathnames
were chosen under the assumption that only the JSON
generation path is being tracked (no information on which
path in the name). This tests compilation on fuchsia
which might have different performance than on host, but
it should be similar.

Initial value:
fuchsia.fidlc_microbenchmarks: Struct/16  585483 +/- 3367 ns

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/440055
Original-Revision: 52a7f3b847b1576a96ba9ac5fe05fe66ff5111c6
GitOrigin-RevId: 3e85c0dac2c030cc75d5d953adc3282de1d85fc4
Change-Id: Ib5bf0cd328653bfc21075b28d1ea899a37237972