[roll] Roll fuchsia [bt][rfcomm] Add support for establishing outgoing RFCOMM channels
Adds a new method to initiate channel establishment.

In general, when opening a remote channel, we must 1) Make sure the
session multiplexer is started, 2) Negotiate the parameters for the DLC
(and transitively the session) by sending a PN Mux Command and 3) Open
the channel by sending an SABM command over the user DLCI.

Once the remote peer positively acknowledges with a UA response, we can
open and establish the channel and relay it to the profile client.

Original-Bug: 59585
Test: Adds a test that verifies the complete open remote channel path to
bt-rfcomm-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/432495
Original-Revision: a24f88f52452736a2d66d41c7ef17bfb79974869
GitOrigin-RevId: 29ba66f699bda4fd463f2439b196249df29bf20c
Change-Id: I793a707fe96b18988b247f083be67c49e251586a