[roll] Roll fuchsia [fidl][hlcpp] Flexible bits support
This change is very similar to the one for LLCPP i.e.
8dfdc60b554992c130ed50e027366d39962846e7. I largely copied the bits
golang templates and unit tests from LLCPP. The main differences are:

- HLCPP strict bits need to stay as a `enum class` for now, because of
breaking changes otherwise. The missing APIs from strict bits were
removed from the unit tests.

- Removed `AllowingUnknown` in favor of `static_cast`, and we intend to
do the same for LLCPP.

- Renamed `MyBits::mask` to `MyBits::kMask` to follow C++ style guide,
and we intend to do the same for LLCPP.

In order to support the bits API under C++14, language version macros
are used to conditionally define the `constexpr` member variables in the
header when under C++17, and in the implementation file when under
C++14.

Original-Fixed: 61395

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/441379
Original-Revision: 28f47529ab47d42bf5daac9b2924f2d45fc7f46a
GitOrigin-RevId: d00dfcca310cf42bb4a66a9b4760df8bdad64807
Change-Id: Id43b0904285317c0c58a3cf64c5e2e11f4e0984a