[roll] Roll infra/recipes [checkout] Skip patching infra/recipes by default
This allows Fuchsia checkouts in general to avoid attempting to patch
infra/recipes, which is an error. The only case where this is valid is
when we're running recipes/recipes.py, so an exception is granted
there.

This will enable scheduling of recipe-versioned tryjobs which have an
infra/recipes GerritChange without erroring out during patching.

Original-Bug: 58899
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/440474
Original-Revision: 52e84c5acc288c17668827ce5f9e93efe8bb0663
GitOrigin-RevId: 3c8358a711210e85f7ccc52ffe741c22afb72df8
Change-Id: I7e58569ee5145ce0035a6c9560cd9348f4a0efd8