[roll] Roll fuchsia [camera][device] Delay serving outgoing directory
The device watcher tries to connect to the camera device as soon as its
outgoing directory is exposed. If we are unlucky, this will be before
the outgoing service has been added to it, and the connection request
will be dropped.
Instead, we delay serving the outgoing directory until the service has
been added.

Test: camera now always works on first connection after reboot
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/440956
Original-Revision: 00661887fe9fb5d5423b63bf26c3a2847202bb9d
GitOrigin-RevId: 2211424ce2bfeb889fa1fa2c50c8e0869b64030d
Change-Id: Ic7a40a9713ed0f26b3690ecc3faa7ca189269130