[roll] Roll fuchsia [fvm] Support reserving slices in fvm host tool
Reserved slices are held in a dummy partition with a particular GUID.
This partition can be arbitrarily reclaimed by FVM later for its own
purposes.

The dummy partition will be hidden by the FVM driver; that is, no block
device is bound for the dummy partition.

This is necessary to reserve slices for the upcoming A/B snapshot
effort, which requires some extra metadata.

Adds a new flag to the sparse format marking a partition as an internal
reservation partition, since the sparse format doesn't have GUIDs. The
flag is translated to the appropriate GUID in fvm_sparse_image_reader.
Note that fvm_sparse_image (which generates sparse images) is *not*
updated in this patch, which is okay because that code is only used by
tests for now.

This change does not include new unit tests for fvm-host, the change was
manually tested. The plan going forward for testability is to verify the
migration code in the FVM driver which detects this reserve partition
and takes over the slices for other uses, once that migration code is
implemented.

Original-Bug: 62062
Tested: Booted on a variety of devices.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/440062
Original-Revision: 664bc86a75d263ecd4fa016c8417dfe0dd523bb3
GitOrigin-RevId: eb1ad90f06039ea86034709875e161065715c2f5
Change-Id: I8e7f9c7a5949f553332ebc092c5b91060356b27c