[roll] Roll fuchsia [fidl][rust] Populate next handle metadata on the encoder / decoder
This CL makes information on expected handle rights and subtype
available when it is time to encode or decode a handle. In
follow up CLs, the encoder and decoder will switch to use
HandleDisposition and HandleInfo and validate / encode rights.

Because there is no structure representing a "handle site",
this metadata needs to instead be generated on the struct,
tables and unions that contain handles. In a previous design,
this information would be passed through encode() calls until
a handle is reached, but for simplicity and likely performance
changes from the extra parameter the metadata is instead
attached to the Encoder and Decoder.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/439600
Original-Revision: 08f152943da7838729fef71401b21cc58b1379ee
GitOrigin-RevId: b20976f0dde727b8a9778bcc84915b11fc794a07
Change-Id: I471fcb4910f10e783a5680f6b1d1cb3e6109775c