[roll] Roll fuchsia [Build] Clean up update-to-zedboot.
This is part of the effort to remove zedboot from build that does
not use zedboot as recovery. According to etryzelaar@,
'update-to-zedboot' is no longer required so remove this.

I also don't see a usage of 'update.sources.manifest' so remove
that as well.

Tested: Presubmit.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/441721
Original-Revision: d5e71851680e9db5511a18d438de3d09d3932175
GitOrigin-RevId: 69a614818c38ebb677719e0d7418d5e69130801e
Change-Id: Iea83a856b9ee977f6c4ab3d4807b2a949b7424cc