[roll] Roll infra/recipes [build] Don't read multipliers if triggering project not checked out
If we fail to find the triggering project in the checkout when trying
read multipliers from the commit message, we should not return any
multipliers. Previously, we instead would fall back to extracting
multipliers from the git repo in the root of the checkout, which led to
incorrectly multiplying tests on CLs in non-root repos where the latest
commit in the root repo had multipliers.

Also, convert all manifest remote URLs to https scheme when checking for
equality with the triggering project's Gerrit host, to ensure we don't
miss the triggering repo if it uses an sso remote.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/441539
Original-Revision: 84948e58916e1d473e4fb43daeda21eccc932c5d
GitOrigin-RevId: 25985eba277717fd59e2e8c688f25a2fe63e21c3
Change-Id: I86cb4165f198435c1103f342f64a163548fbee0c