[roll] Roll fuchsia [ubsan] Disable vptr checks on host to unblock rollout
While the right thing to do here is to have up-to-date LLVM prebuilts
with variants, that is currently likely to take a while.  We can get
most of the upside of turning on UBSan on host tools by disabling this
particular check on host (where we do not have any UBSan automation
today) and following up.

Original-Bug: 60446
Original-Bug: 60253
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/441756
Original-Revision: 6a1096bdffdddcf758e1facba84e77a7cdf86e31
GitOrigin-RevId: 7551ce711acc4b9df8e853fef04881c94ef72cf5
Change-Id: Ica7b6f7b2c98e3c7a19f7cad61306a8a2dbac288