[roll] Roll fuchsia [settings] Add future input FIDL format
This adds the new, more generic interface for input. It allows for
different hardware and software states and adds a camera as a device
type. After this, clients will have to be ported to the new api, then
the old api can be removed.

Original-Fixed: 60676
Test: fx test -o setui_service_tests
LSC: Just adding a transitional FIDL, which prevents it from breaking
rigid tests. A full LSC process will be run when clients are switched
over to the new API

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/440279
Original-Revision: fb51932d9394ed59888eccebc9c085244df1d71f
GitOrigin-RevId: 365e1efb266901e4dba864421771c967435561db
Change-Id: I2e10b1222d8ea24f3772282a2faad7cf58c20a43