[roll] Roll fuchsia [hid-buttons] Refactor unit-tests
Refactor the unit test so that buttons are set up based on
their own configuration and not what type of test is being
run. This removes a lot of boilerplate but runs the test
in an identical way. It should make it easier to add
more tests in the future.

Original-Bug: 62286
TEST: fx test hid-buttons-test

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/442061
Original-Revision: 44ef772f9a5b3f1375ce4503cbc924b8f3496940
GitOrigin-RevId: b0ffa065a81ae9fbe77ac272e90414eb4503dfe4
Change-Id: I37eb09e1d9d7fa50b22ace13dc1435388593b68a