[roll] Roll fuchsia [web_runner_tests] Remove use_legacy_observer_behavior
It is an obsolete workaround to allow transition to the new behavior.

This is a manual revert of most of http://fxrev.dev/282890, which was
before the file was moved from topaz. The updated new behavior (when
use_legacy_observer_behavior is false) is retained.

Original-Bug: 29937

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/441853
Original-Revision: 638b7b3a6940c40ea5f818231c272b00ff5d9e73
GitOrigin-RevId: ed33030e7302e785627849279d2f78122634608e
Change-Id: I0b4dd4a374a69fd918609d0d19ce8f5199dd08e2