[roll] Roll fuchsia [overnet] Fix recursive lock
with_handle() locks a handle table and calls a function - during the
review of this code I missed the fact that we did this recursively,
which will at some point trigger a deadlock.

This change adds some debugging machinery that would have caused a test
failure previously, and a fix for that.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/441039
Original-Revision: 9798e5fb2dbb24618bb2390ae5d536bf743c874e
GitOrigin-RevId: 0670d4ad51a2e4ba0484c94a8230a00f2d7b1fc4
Change-Id: Ib9817f9516ac64a4e6fe6a8df096fd07d0b6b17f