[roll] Roll fuchsia [fint] Add GN rule to generate .pb.go files
And update .pb.go files with the files generated by this new rule.

Add a script to update them and remove the old gen.go.

The diffs are due to using a different version of protoc and/or
protoc-gen-go.

Original-Bug: 61402
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/442057
Original-Revision: 6f06dad9eaf7435bb5ff818d9330ae347eaea3c6
GitOrigin-RevId: a111505127da4e0db309c2d443daad8fd42ec7b7
Change-Id: Ie28b52be6e580da9c763de9768f3fb92c17b1c71