[roll] Roll fuchsia [kernel][vdso] Sync CPU startup during VDSO creation.
Make sure to wait until all CPUs have started on ARM before deciding
whether or not we need to install the A73 system timer read
mitigation.  Failure to do this would mean that we have a race, and we
might end up not installing the hook when we really need to do so.

Don't wait forever to make this decision.  If something goes terribly
wrong with CPU enumeration, simply install the mitigated read and take
the penalty whether we know that we need it or not.

Original-Fixed: 59609
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/425313
Original-Revision: fa0b458f83b73f44ec24297883d6a07cecc9be1b
GitOrigin-RevId: f038c93f12edd30cfa1e83d7a1d6b94d638e3cdb
Change-Id: Icf5c81bbc2ddbf0159cbd8eda3616b194dd7b438