[roll] Roll fuchsia [fvm] Ensure reserved partition is paved
The reserve partition added in fxrev.dev/440062 was always being deleted
during paving, since it started out as inactive and was never made
active by a call to Upgrade.

Fix this by making the paver always upgrade the reserve partition.

Doing this required exposing the previously-hidden reserve partition as
a regular block device, since the paver makes a FIDL call to the block
device to get the partition GUID before attempting to upgrade the
reserve partition.

We also change how we detect internal reservation partitions to matching
the type, not the GUID. This is necessary because the GUID changes every
re-pave, but the type is stable. Note that this detection heuristic is
currently only used for logging, but eventually we need it to work so
that we can reclaim these slices.

Tested: Paved from before->after this change.
Tested: Paved from this change->new build with this change.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/442196
Original-Revision: 164cf59383e03eab8b75a0cd03d5bfa3040fbc19
GitOrigin-RevId: ce206838ee74775941ae765e906f67612248d572
Change-Id: I605697588553e3c9bfc17731d1ede96f4f19ec1b