[roll] Roll fuchsia [build] Add `enable_variants` parameter to toolchain_variant_suite()
This adds a new parameter to toolchain_variant_suite() that allows
the caller to enable specific named variants, evern if they are not
listed in the global 'select_variant' global list.

This is necessary to build Asan-enabled variants of the C library
in the Fuchsia build, a requirement to move the Fuchsia sysroot
generation from the Zircon build.

+ Fix sanitizer_default_options.gni to properly handle Zircon-specific
  toolchains in the Fuchsia build.

Note that this only ensures that the GN toolchain() instances exist,
but does not add any dependencies to the build graph. For testing,
add the following to your universal_package_labels list:

  "//zircon/system/ulib/c(//zircon/system/ulib/c:user.libc_x64-asan)",
  "//zircon/system/ulib/c(//zircon/system/ulib/c:user.libc_x64-asan-ubsan)",

Then one can build the following targets with 'fx build':

  user.libc_x64-asan/libc.so
  user.libc_x64-asan-ubsan/libc.so

Original-Fixed: 62555
Testing: See above.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/441821
Original-Revision: ccb0996b14fef8af0605a22414bc5041a5d3adfa
GitOrigin-RevId: 5a8b1014a149eff7693cefdc4e8d34e3415a8a8b
Change-Id: I244b6e1df4abcea445d156bdbde7855ec11599c2