[roll] Roll fuchsia [fshost] Add options to fshost configs.
Previously the fshost build files were checked in and added as resources.
This change generates the build files when GN runs.

Adds a "default" set of options for the normal device target for the
urrent build. With the current settings this will contain only "default"
which will have no behavior impact. This changes the storage of the
options to no longer store "default" in the option set (this is handled
at a higher-level by the decoder).

Adds options for the max minfs and blobfs partition sizes. These are not
currently hooked up. Support is added for key/value options in the
fshost configuration.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/441923
Original-Revision: 13337cb507382ac33526c0f1c2b5428768d68bac
GitOrigin-RevId: 0a5519a31f065540c58799fe842854d98a990b1f
Change-Id: I4ee15ab3e41ef941300c29300480d960f7308331