[roll] Roll fuchsia [component_index] limit maximum number of results returned
On a build with a large number of packages (such as a build
from `fx set workstation.x64 --with=//bundles:tests`), the
number of components matching a query may exceed the
maximum size of a Zircon message. That causes the
component_index service to crash.

Fix this by setting a hard upper bound on the number of
matching items returned. Note that it's still possible for
the service to crash if the sum of the lengths of the
matching component names is large.

However, that problem is only hypothetical for now. So
we'll leave it for another day.

Original-Bug: 45097
Test: fx test component_index_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/442379
Original-Revision: ca098ef7c6f91b8338bd52cb5d581ad337d4633c
GitOrigin-RevId: fbb4a4f772bc66bba474ae7bba68f30fe593cafc
Change-Id: I3da13dd44acdf548a12367e9198402f9d8cc22d7