[roll] Roll infra/recipes [auto_roller] Ignore merged identical CLs
If a roll CL has an identical diff (and thus an identical Change-Id) to
an already-merged CL, we should ignore that already-merged CL and create
a new CL instead of aborting, because it's possible that a roller would
create identical CLs at two different points in time, e.g. in case of a
revert and subsequent reland.

However, if we detect an identical CL that is *not* merged, we should
still abort the roll attempt to avoid a race condition with that CL. I
improved the build UI in this case to explain why the roller is aborting
instead of exiting with a cryptic "OPEN != ABANDONED" message.

Original-Fixed: 62069
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/442524
Original-Revision: e40828bfdc94f18b8667de97c56b48963b34db5d
GitOrigin-RevId: 75694b5895b666673958f25e3459762d042e51d0
Change-Id: I827d28bfc270b02e6740ef34e22f287bc65756ac