[roll] Roll fuchsia [fidl][sdk] Remove fidl-async from the SDK
//zircon/system/ulib/fidl-async contains a basic message dispatcher
implementation for the C bindings and LLCPP bindings. The library was
only included into the SDK because the HLCPP library declared it as a
dependency, which it doesn't need. Since we are deprecating
`fidl-async` and both C and LLCPP bindings are not intended to be used
from the SDK today, we should remove this library from the SDK by
removing the HLCPP dependency.

I did not find mentions of `fidl-async` / `fidl_async` in any partner
repos.

TEST: LSC presubmits

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/442018
Original-Revision: 0fc4c5896a6cdd9d2f6a4878a484427e4a189c46
GitOrigin-RevId: c36faab71d32fab4fc02b79976ed5e7062a2f450
Change-Id: I1f53ae359a7f1beaf24ff869f56a5bb2ed4804f6