[roll] Roll fuchsia [wlan][policy] Exit when starting AP fails.
This change eliminates log spam in the event that starting
the AP fails.  Currently, a failure to start is
irrecoverable and the SME proxy is useless.  Exiting the
state machine allows the policy service and its clients the
opportunity to cleanup the interface and try again.

Original-Bug: 61097
Test: fx test wlancfg-tests
Multiply: wlancfg-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/433655
Original-Revision: fa59a12fa27ffc17853e7966673d342fce681c84
GitOrigin-RevId: 3a44aa2ffd502758b98e6e8ab40f76767efd6c38
Change-Id: Ic64b6a6b762d4aac3555125b54e39490da2133fd