[roll] Roll fuchsia [fuchsia-audio-codec] Fix off by one check on stream write
Fix check for space available for the case when space_left is equal to
the amount of bytes left in the current input buffer. This allows input
packets to be sent when exactly filled rather than waiting for the next
data to come in.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/433873
Original-Revision: 16b4d506e559dabf7725e629e9961bfa90cae6e0
GitOrigin-RevId: 5d40c0aa1890d32e87212eceab55cabb26c8e4c5
Change-Id: Ia03e76a52905d6af419f57facb82c66064e2644d