[roll] Roll fuchsia [gn sdk] Remove bash -i from fserve.sh script to fix unexpected side effects
The #!/bin/bash -i was added in fxr/410994 and resulted in some
unexpected side effects. For example, femu.sh on OSX would suspend
itself on startup, somehow due to the -i creating a login shell that
changed the terminal properties. There were other problems with error
messages appearing in the output, although this was fixed in fxr/414742.
I suspect there were other bugs introduced by this as well.

However, the bash -i was originally introduced to fix a problem where
gsutils was not being found if in a non-system PATH location. Users will
need to add gsutils to a system PATH directory, and cannot rely on a custom
.bashrc or .bash_profile setting.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/440181
Original-Revision: df87cea134cc656d72e1c5c49a8f68199e4be179
GitOrigin-RevId: 3385ba3baa70f55e502858406d4d52b1b3e8a6ed
Change-Id: I56dd04f6823c70469aa9c490faa3bb26c11e492b