[roll] Roll fuchsia [mock-function] Use ZX_ASSERT for checking the number of expectations
This causes an abort with the following message:
ASSERT FAILED at (.../lib/mock-function/mock-function.h:190):
    enough_expectations_were_set

ASSERT_LT just returns if not enough expectations are set, meaning a
crash will eventually occur when trying to dereference the null
expectation. Switch to ZX_ASSERT so that execution stops when the fatal
condition is detected, and so that a slightly more helpful message can
be printed.

Test: runtests -n mock-function-test
Test: Deliberately called a MockFunction with no expectations, saw the
      above abort message
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/439203
Original-Revision: d97260729f32e300f3327afbc8e2a50371233603
GitOrigin-RevId: 511d23af6434efecbc443b05750a9b454d174b63
Change-Id: I2a02f2594bc8ee78f1d93b41a3b2bf582a590049