[roll] Roll fuchsia [bt][a2dp-sink] Use Codec Negotiation
Use Codec Negotiation to select streams when we are initiating
streaming.  This should mean that we do not just select 44.1khz
which is not required to be supported by all sources.

It also means we select AAC when it is available in the source.

Original-Fixed: 39321
Original-Fixed: 48529
Test: added connect_initiation_uses_negotiation in ConnectedPeers
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/442154
Original-Revision: 399ace7de2851e0fff15649d1ad7fc78b7decd52
GitOrigin-RevId: 566792437e2ddf3d0836491a713bc3cd54366d81
Change-Id: I7db7b2a45c81f1868190091c2e9ac840ebfbc1ac