[roll] Roll fuchsia [bt][a2dp-source] Fix panic in encoding task
When the encoding AsyncWrite only had enough room for part of a buffer,
the encoding stream process would mark in the buffer where to resume,
push the buffer back on the queue, and then return to the top of the
loop to re-poll to register the waker.

The top of the loop would pop off the current buffer, and drop it on the
floor when it got a Poll::Pending.

This would also happen whenever the AsyncWrite for the encoder was
stalled, so we were also just occassionally dropping packets when the
encoder could take no packets.

Retain the buffer when we get Pending from the AsyncWrite.
Make the EncodedStream a bit more generic, and add a build_test which
doesn't require a StreamProcessor to operate the stream, making it
testable.

Original-Fixed: fxbug.dev/62547
Test: added test_stalled_encoder_input in bt-a2dp-source-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/442996
Original-Revision: 398f8811bb8a2f266e9c28090fc46b8d63628371
GitOrigin-RevId: e2f4ae15cd6f3bcd9ce55a15d3693ec069919bf9
Change-Id: I5606c6bf66b61a6e53afa31e6d5c827d9623f5f6