[roll] Roll fuchsia [hid-buttons] Refactor notifiers
Update the notifiers so that they rely on ButtonTypes
and are mapped by ButtonTypes. The notifiers are
listening to ButtonTypes via the FIDL API, so they
shouldn't be organized by BUTTON_IDS_

This has the added benefit of using the type system
more to make it clearer which ID's belong to what.

Original-Bug: 36834
TEST: fx test hid-buttons-test

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/442199
Original-Revision: 93a56cb0eec0cefcb513b3b039a6ea20a2e12171
GitOrigin-RevId: c0d3d8d3dc7194dcf5a8802b8c99cc3dd8c7a7c0
Change-Id: Ia21cf5a180164a5edcc7f0f78647065ac242ff44