[roll] Roll fuchsia [system-updater] Add a timeout to RebootController
As a precaution against misbehaving implementations of
fuchsia.update.Manager, this change adds an unreasonably long timeout on
waiting for the initiator of an update to signal that it is a good time
to reboot.

Also, this change reworks the RebootController to own the async task
waiting for the signal, which ties the lifecycle of the fidl protocol
handler for the reboot controller to the update attempt instead of the
client's connection, preventing misbehaving clients from keeping async
tasks around in the system-updater.

Alsoer, always provide a RebootController to the update attempt in the
fidl path, simplifying removing that Option<> later.

Original-Fixed: 61956
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/442686
Original-Revision: ff7a56c5dbb3ea4cfa5fb9209c2fe1b378021c98
GitOrigin-RevId: 5c4d110c25c07591c507a794f48724c55f910dfd
Change-Id: Id91061d66c5e4a78ecf310dfc2399bcac45c3ba1