[roll] Roll fuchsia [ffx] Add default config rules for client commands.
This change does a few things:
-- Makes `-t` set `target.default` in the runtime config.
-- Makes a macro that allows for config fallback of any arbitraty struct
   field that is an `Option<String>` and `Option<f64>`
   (more types will be added as needed).
-- Changes get-ssh-address command to use `target.default` rather than
   accepting a parameter.
-- Uses backup of timeout to a config value.
-- Removes use of `inner_lock` on TargetCollection to make deadlocks
less likely to occur from misuse (as there is now no way to access the
inner locking mechanisms).
-- Removes timeouts inside the daemon, allowing for target queries to
run an arbitrary amount of time if necessary (as long as the client
stays up).

Example usage of macro:

```rust
 #[derive(FfxConfigBacked, Default)]
 struct Foo {
   #[ffx_config_default(key = "timeout", default = "1.0")]
   field_one: Option<f64>,

   #[ffx_config_default("nodename")]
   field_two: Option<String>
 }
```

These will generate according functions for looking up the respective
values from the FFX config.

This is used in the two plugins, one for getting the default target, the
other for setting a timeout for getting the ssh address of a target.

Original-Bug: 59622

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/436334
Original-Revision: 61c5dcbfcb856a7f3efbd6eff81460d7eec0d9ea
GitOrigin-RevId: 5ddd62f9b538ceae7fe2a23cb73927997f9da0b9
Change-Id: I7e234c75cfa4235afc0f9a83225650b5bf151543