[roll] Roll external/github.com/catapult-project/catapult Reduce warning output in tests
When running minidump tests on Windows there are some warnings that are
printed every or almost every time. It is not obvious to somebody
looking at the test output that these warnings are known and expected so
they can lead to wasted time. This change suppresses three of them:

- 'Failed to import devil.android.forwarder' - this cannot import on
Windows and printing this message adds no value.

- 'Unable to stop ts_proxy_server gracefully' - signal.SIGINT is not
supported on Windows so this scary looking failure message will be
printed every time, which adds no value.

- UnicodeDecodeError: 'ascii' codec can't decode byte 0x89 in position 80: ordinal not in range(128)
This error appears to be a cross-platform bug that happens when we try
to long .dmp and .png files as text. This issue is further confused by
the fact that the .dmp file extension is used both for crash dumps and
for the textual symbolization of those crash dumps. The fix is to not do
the logging for .dmp and .png files.

Presumably these warnings affect many other tests as well. These changes
should be safe for other tests, but may be incomplete.

Original-Bug: chromium:1084931
Original-Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2496672
Auto-Submit: Bruce Dawson <brucedawson@chromium.org>
Original-Revision: 1332bc3b15d4461b4490d1a615717f4d5b347447
GitOrigin-RevId: 425abccd233757444cf9a8e3269d95f3f5c19899
Change-Id: I00f3e3a1e6581423d52073f994688b4be9533801