[roll] Roll fuchsia [wlan][policy] Expose more info from scan results
This CL's primary purpose is to route the information "did we observe
this network in a passive scan" from Network Selection through to the
Client State Machine, which (on successful connection) will record that
information in the Saved Networks Manager.

This information will help improve our determination of whether a
network is hidden or not.

The two main changes to enable this are:
1. within the Network Selection module, keep all the scanned bss
   information (rather than simplifying to only a single RSSI).
2. include a "network metadata" parameter in the ConnectRequest,
   which contains the "network observed in a passive scan" boolean.

Original-Bug: 35918
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/441855
Original-Revision: 06ac3fed9b319352a94e3ba8f229f0ff8d1febfb
GitOrigin-RevId: 1c6bf178997d7896a284626cfe96e66436111b05
Change-Id: I89a0d3103f6b43098772636d515e0e024e5a5d64