[roll] Roll infra/recipes [build] Simplify zircon_extra_args setting
Instead of conditionally initializing `zircon_extra_args` to an empty
scope if is hasn't been set every time that we want to set a value in
that scope, we can have the recipes initialize it to an empty scope
before setting any other args (but still conditionally and after all
imports, because imported files might set it). This will make it safe
for other args set by the recipe, as well as args set in the configs, to
assume that `zircon_extra_args` is already set, making it less verbose
to set `zircon_extra_args` values: just do `zircon_extra_args.foo =
"bar", with no need for initialization.

Some parts of the configs assume that `zircon_extra_args` is *not*
initialized by the recipe, so this will require a soft transition with
the configs:
1. Add `if (!defined(zircon_extra_args)) { zircon_extra_args = { } }`
   before every `zircon_extra_args` value set in the configs
   (http://go/fuchsia-internal-review/272242).
2. Land this CL.
3. Undo #1, and remove all `defined(zircon_extra_args)` checks from the
   configs (http://go/fuchsia-internal-review/272243).

This will also require a corresponding change in fint, although that
won't require a soft transition because fint isn't yet used on any
builders that set any `zircon_extra_args`.

Original-Bug: 61977
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/439695
Original-Revision: 73f214434093215316474cf9a2f62eb28a80f62d
GitOrigin-RevId: 6a4d14c95145e1d9ef7bdba64f300bb6ac2a448d
Change-Id: If0d05230cccfafa7fd7529c3ac770db7614fb8ce