[roll] Roll topaz [topaz] Fix race condition in platform view creation
This change is trying to address http://fxbug.dev/60199 where
the flutter-embedder-test is flaking. The crash happens when the
embedder is accessing the view_holder in UpdateView, before
CreateView has been called.

The fix is to wait for the CreateView to complete before the child
view is rendered. CreateView is called by the
ChildViewConnection::Connect.

Original-Bug: 60199
Test: flutter-embedder-test passes
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/topaz/+/443294
Original-Revision: 5533d70bc8b7eab4a21fc3513e035236fd26d156
GitOrigin-RevId: bf1f5471a77018030575f1a8b780daa0fd349e29
Change-Id: I8a2002614bb7b108bd34082cc1eb3da05c6798e0