[roll] Roll fuchsia [fvm] Add oldest_revision to FVM header.
This implements the policy in //src/storage/docs/versioning.md for
tracking the oldest revision of the software that has touched the data.
The current revision is set to 1 so we can differentiate headers
produced before this change. The format is unchanged and is
backwards-compatible with old versions of the driver.

The existing "version" field is renamed "format_version" to align with
the document's terminology but its meaning and binary representation is
unchanged.

Update FVM driver test to write a real FVM header to the udnerlying fake
block device and initialize the FVM driver like it would be in normal
operation. This allows removing some test only code, allows testing of
Load(), and makes the tests generally more realistic.

Adds tests for oldest_revision updates.

Original-Bug: 62594

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/442564
Original-Revision: 5c534a9423e5cf261a0bff42972077e9b5487e8b
GitOrigin-RevId: 261261c65f1f2f4e3498cfea2038e46e41c77acc
Change-Id: I6a4d43c5200141617baeca43b5a227d55b092f91