[roll] Roll fuchsia [e2e][intl] Stops preemtively requesting ViewProvider
The test used to always requst ViewProvider, as a way to jumpstart the
Dart program when running under Flutter.  Old code did that anyways for
Dart as well, even if unnecessary, which proved to be flaky.

New code removes requesting ViewProvider when unnecessary, which should
reduce this source of flakes.

Original-Bug: 62736
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/443267
Original-Revision: 4ad0d5717d65736856bdc55819a4cebc258db76d
GitOrigin-RevId: c9bf56583166b6da7abacb4f7d363c69015d7e88
Change-Id: I23bd24afac57cc1635c5db49332fc185c220f056