[roll] Roll fuchsia [blobfs][chunked-compress] Add Simplified DecompressFrame.
Adds a method to chunked-compression for doing a simplified call here
the caller is responsible for ensuring that the start points and lengths
encapsulate a complete frame rather than using the seek table to
describe them.

This is used in the remote decompressor where the remote process has no
information about the seek tables and performs the decompression relying
on blobfs to give it the correct values.

BUG: 62396
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/442628
Original-Revision: 4dc06e40cadcd7fdc901cd38451934d519584b04
GitOrigin-RevId: b5ec40fe41ed95fcee99a3755d130e7ed8db5098
Change-Id: Ifaff5565a68c230bbb4e84784e3e365907ef0b32