[roll] Roll fuchsia [scrutiny] Implement Config options for Scrutiny.
This change formalizes a public interface for Scrutiny runtime
configuration. Prior to this change is a third party wanted to integrate
scrutiny into their project they would have to pass in command line
options manually. This leads to several issues:

1. We may want specific runtime configurations that we don't want to
   expose as command line startup options.
2. We want to be able to have a codeable configuration that can be
   configured directly in the code.
3. We want to provide sensible defaults for different integration
   strategies.

This change address this in the same way we do this for the
component_framework and other areas of the code base with a basic nested
structure config. This CL doesn't introduce configuration serialization
for simplicity.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/442735
Original-Revision: c1adf4e2cd083a527eff12bf2611089caae71a6e
GitOrigin-RevId: a93e15db6540532664b557b00f0358518c4ab613
Change-Id: I626bba8df2c0558e33fc073d79c78c6b318554ad