[roll] Roll topaz [dart][fidl] Support unions with escaped names or members
If a FIDL identifier conflicts with a Dart keyword or other reserved
Dart identifier it will have a $ appended. This is fine until we
generate a string that contains that name - which is interpreted as
string interpolation.

This fixes the case for unions with escaped names and member names.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/topaz/+/443755
Original-Revision: a880788f8ebe20dbba414482bcb4631cafa695c5
GitOrigin-RevId: 85a379b5ae86111c48fc333397f2ecd22c888948
Change-Id: Ib7b6cc2d1f9f0e702eeba8ba1b2b9be0d47863b4