[roll] Roll fuchsia [kernel][vm] Introduce VM hierarchy base type
This moves the references to the VMO hierarchy state as well as methods
for retriving the locks to a base type. The only actual effect this has
is that the memory layout of the VmObject struct is possibly different,
but otherwise there are no semantic changes.

Motivation for this is to prepare the way for additional objects to
be part of the VMO hierarchy and share the lock and other state,
without having to be based on VmObject.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/437340
Original-Revision: 4175378c02d63cb41d88019c01e2345ce4891651
GitOrigin-RevId: 69fe6a8d65a21fe8d4b8cf0ca4c73054fac12229
Change-Id: I8a5b27a15e9be0d08a76745d5dea8418c24e0a97