[roll] Roll fuchsia [kernel][vm] Introduce VmCowPages
This creates a parallel hierarchy to VmObjectPaged in VmCowPages. The
logic of actually manipulating copy-on-write through the VmPageList is
handled by VmCowPages, with VmObjectPaged maintaining responsibility
for handling VmMappings and providing the user API.

Standalone this commit is a temporary regression in terms of memory use
and performance due to the VmObjectPaged still creating needless hidden
parents that serve no purpose. The reason to have a temporary regression
is to keep this commit as close to "moving code" as possible without
there also being additional complex changes hidden inside. This does
lay the groundwork for successive commits to perform the 'interesting'
transformations to end up in an eventually more efficient hierarchy.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/437341
Original-Revision: 08ec46b68186331f5275098ff33fca66dc5e1b36
GitOrigin-RevId: 2c11c462ddf8799733cf2e888454056dfa28314d
Change-Id: I02934a58460f4d7de966d1b36aabe69e3ca54acc