[roll] Roll fuchsia [kernel][vm] Hold lock over VmCowPages::Commit
This redistributes the commit logic between VmCowPages and
VmObjectPaged. The goal is to move the top level retry loop that has to
wait on page requests with the lock dropped in VmObjectPaged, and have
the VmCowPages perform only the section that can be done with the lock
held.

Motivation for this is that later we would like the cow_pages_ link
in VmObjectPaged to be modifiable to point to different VmCowPages. If
VmCowPages would be dropping the lock mid operation then it could
potentially cease being the linked object, and  when the lock is
reacquired be acting on the wrong VmCowPages. Dropping the lock in
VmObjectPaged ensures that after we re-acquire we will use any
potentially updated cow_pages_ link.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/437345
Original-Revision: 790cd5ce767200a0269d5bb0e550c70b23167404
GitOrigin-RevId: f2a0fd0ad0bf6102b87f55b368a0385d76f98222
Change-Id: I081924873ac2268327411eeb4c2c7baf503abe8a