[roll] Roll fuchsia [kernel][vm] Remove hidden parents from VmObjectPaged
This removes the notion of hidden parents from VmObjectPaged and in
the process places control of the VmCowPages hierarchy manipulation in
the VmCowPages code, instead of being done from the VmObjectPaged.

Aside from the removal of the redundant VmObjectPaged hidden parents
this also performs an additional optimization of making the parent_
reference in VmObjectPaged a raw reference and not a RefPtr. This
can be done since now that the copy-on-write pages are held in
VmCowPages there is no reason to hold the VmObjectPaged around after
any other references, such as user handles and mappings, have gone
away. Not holding the parent via a refptr also ensures we can not have
any chained deletions and so the deferred deletion mechanism is only
needed now by VmCowPages.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/437347
Original-Revision: 6770011d81eb6bca3b33c169da02be56c2773ceb
GitOrigin-RevId: d9422fe6c921271f6022c511c830f066b80d1f0e
Change-Id: I6dc77cc0b6b6942cf97a66190d7f0d9ea8f90b37