[roll] Roll fuchsia [dhcpd] Use ClientIdentifier instead of chaddr
Formerly, dhcpd used the client's hardware address as a client
identifier. The DHCP protocol provides a ClientIdentifier option which
should be used for this purpose; the ClientIdentifier option is a client
provided opaque byte string. This change replaces use of the chaddr
field as the client identifier with the byte vector in a
ClientIdentifier option.

Towards that end:
- replace MacAddr with ClientIdentifier as the key for stored client
  configuration
- add From<&Message> impl for ClientIdentifier
- encode the ClientIdentifier as a two-field ("variant:hex") string
  where a string representation is needed, e.g. error messages and as a
  fuchsia.stash key

Original-Fixed: 20914
Tested: fx test dhcp-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/440981
Original-Revision: ffa9f614bf6fe800cf561fe0e32954a314558256
GitOrigin-RevId: 0426cd98a958b7010607e311c25dd5ea9e9cc970
Change-Id: Id5f9fae5d2fe1b135707b52f51b94a8a3dd2af5c