[roll] Roll fuchsia [jitterentropy] Break dependency on arch/x86/feature.h for InvTSC bit
Jitterentropy is a third party C library linked into the kernel to serve
as an entropy source for the kernel csprng. Unlike much of the rest of
the kernel, it is compiled as C source.

On x86 systems, it tests for an Invariant TSC, to decide how to
incorporate timing information. Convert its use of x86_feature_test to a
raw call to CPUID - this allows us to remove the inclusion of zircon
header files, which allows us to make those headers C++-only.

Original-Bug: 47117 Use only ktl::atomic in the kernel

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/441822
Original-Revision: 4e17045be412b189eb27dd283dfb3fae6db085ab
GitOrigin-RevId: b111aedcafd79da85d67061ec6ede2f226c1fd2e
Change-Id: I60f22db48e0b98f7b42bbfe37eb7298977b96589