[roll] Roll cobalt [Cobalt 1.1 privacy] Decode private indices for histogram reports
Adds a function which decodes a private index as an event
vector and histogram bucket index (for per-device histogram
reports) and a second function which decodes a private index
as an event vector, histogram bucket index, and bucket count
(for FleetwideHistograms reports).

Also changes the order of the bucket_index and bucket_count
arguments in the HistogramBucketAndCount{To,From}Index()
to match the most convenient order for the new decoding
function.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/cobalt/+/443215
Original-Revision: 5c7a0c0537f528310f0c147afc42cf52a921cdb1
GitOrigin-RevId: 0ddf95e571cabeab05843c012671963ffbf21058
Change-Id: I2e60bc9953dd5e8e1066b4991fa58c61c16fc2cb