[roll] Roll fuchsia [SDK] Converting fpublish.sh to binary executable
This creates an sdk-tools project under //tools

The tool fpublish.sh is implemented as a bash shell and does not
meet the CLI rubric requirement with respect to implmentation language.

This ports the bash script to Go.

This is a temporary solution until ffx package management is completed.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/434917
Original-Revision: 33744d3901ff1c006e190b2fe858bd8e0084e505
GitOrigin-RevId: e0ef59b8bc89a556ebfeb429716c605f310269a5
Change-Id: I4447cd7a4d507733d1e0589339f03dcfa0dc9bde