[roll] Roll fuchsia [netsvc] Don't output nodename and exit when zircon.nodename is set.
Netsvc would check for the presence of the `zircon.nodename` boot
argument and set the '--nodename' flag if it was present. This isn't the
correct behaviour, as `zircon.nodename` passes the link-local address
nodename (like step-atom-yard-juicy) from the bootloader to the kernel,
while `--nodename` instructs netsvc to output the board's nodename and
exit.

Remove the incorrect logic to fix zedboot on NUC.

Test: fx pave works again on NUC with ToT zedboot.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/443479
Original-Revision: 694edc9d68d5731f506206e2a8fe900cad423850
GitOrigin-RevId: a80a96646f926109c9be1a6e6956e928374a5e26
Change-Id: I054713dd60f3444e868b31ad75c0db6bdb865076