[roll] Roll fuchsia [wlan] Create fake_bss! macro for mocking BssDescription values
The fake_bss! macro is introduced in this change to simplify the way
fake BssDescription types are created for tests. The first argument of
fake_bss! is always the security type desired, and the remaining arguments
override the default fields of the fake BssDescription. The security
type for the fake BssDescription will determine the default values for
the cap, rsne, and vendor_ies fields.

The fake_bss! macro goes a long way to reduce the amount of custom
"fake bss" functions implemented in various places. Before this change
it was common to encounter function calls like

  fake_vht_bss_description()

in test code. This call of course does very little to explain to the
reader what part of the fake BSS is particularly important to the
test. Now, the reader will encounter something like this instead.

  fake_bss!(Open, chan: fidl_common::WlanChan {
      primary: 123,
      cbw: fidl_common::Cbw::Cbw80P80,
      secondary80: 42,
  }),

The latter makes it much more obvious which fields of the fake BSS are
relevant to the test. Here are a couple more examples.

  protected_bss(b"foo".to_vec()) -> fake_bss!(Wpa2, ssid: b"wpa2".to_vec())

  bss(-10, -10, ProtectionCfg::Wpa1) -> fake_bss!(Wpa1, rssi_dbm: -10, rcpi_dbmh: -10)

Original-Fixed: 62886
Test: fx test wlan-common-tests wlan-mlme-tests wlan-sme-tests wlanstack-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/443497
Original-Revision: 4417e323178fb6497054cd44116e4c9522a7a31e
GitOrigin-RevId: 9d69f318a1a45dbaab9a8195eacf641b6f6ff0ac
Change-Id: I3af62935d8765895f4d3467e822a39340e40bc63