[roll] Roll fuchsia [build] Improve sysroot include directory generation.
This CL changes the way headers are copied into the sysroot's include
directory. Before, a simple series of copy() rules were performed,
which would end up copying various headers to $sysroot_dir/include.

However, if a header was removed from the sysroot entries, its old
copy would linger on in the build output directory, seldom resulting
in a broken CL that still builds locally.

This gets rid of the issue by using a custom Python script to copy
all input headers to $sysroot_dir/include, as well as ensure that
any stale outputs in this directory, that are not described by the
input JSON file, are removed automatically.

Note that list of headers are taken from legacy_sysroot-$cpu.json
for now, which miss the dependency on //zircon/vdso:public to generate
some Zircon-specific headers, so for now add this dep explicitly with
comments explaining it will be removed once we switch to parsing the
sysroot_entries.gni content directly.

+ Move the sysroot directory from
  sysroot_toolchain/obj/zircon/public/sysroot/sysroot/ to
  sysroot_toolchain/sysroot/ since the extra sub-paths do not
  bring any benefits here, and shorter paths are easier to work with
  when looking at build commands during failures.

Original-Bug: 60613
Testing: Should not change the build.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/442536
Original-Revision: 0d4aec5dcc7548ea8483607581e734c489dbadf4
GitOrigin-RevId: 04e818fbd874f6e78e21679160008f7a0534f712
Change-Id: I388e1d01542e0d1d7e71a6ff52a225ba319b6205