[roll] Roll infra/recipes [tricium] Don't skip patching recipe changes
We don't plan on running the fuchsia tricium builder in recipes CQ any
time soon (unlike regular fuchsia builders) so the only tricium builder
that will run in recipes CQ is the dedicated tricium-recipes builder.
That builder actually checks out the recipes jiri manifest and patches
in the CL so it can analyze the recipe code, so it's safe (and
necessary) to patch the recipes repo for this builder.

Test: https://ci.chromium.org/swarming/task/4f8f0ec7679d6110
(If you look at the associated CL, the changed files correspond to the
files checked by the recipe.)

Original-Bug: 58899
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/445155
Original-Revision: cf0b031b623d160937ebd7d53fbf1ee74d13cffe
GitOrigin-RevId: 3a4a17f9784ade14d5f7949a355ee3d2fb22d5fa
Change-Id: I2455c219c1f420e9426aed09a622b03c5bdc1d16