[roll] Roll fuchsia [kernel][x86] Move L1D$ flush to before thread-signaled check
When returning from a syscall with an error result we flush the L1D$
to mitigate Spectre V1.

Prior to this change, we performed the error result check and flush
after a debugger had a chance to modify the returning thread's
registers.  This meant there was an opportunity for a debugger to
change an error result to a non-error result and therefore bypass the
L1D$ flush.

This change moves the check and flush earlier in the syscall return
path to before the thread signals are checked.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/443315
Original-Revision: e8bc613e93e3cb13588124695d130a421497831d
GitOrigin-RevId: cac21bdaa5a5a5ec10af2c6d1126427b5fd4158a
Change-Id: Ib67f64ea985a84957f81a0eb2e4c094a88a3af50