[roll] Roll fuchsia [storage][ftl]: More refined counters.
In preparation for Lapis early integration added counters for the
nested operation of each type based on the input operation.

Noted that the most signficant changes will be on write and flush path,
read may also incur in moves(read perturbation protection) and so can
flush when making space for new map pages.

It will also make it easier to spot odd things happening.

Deleted ftl_proxy, since it was a workaround for lack of cobalt in
drivers, but the recommended workaround was considered an overkill.
(*shrugs*).

TEST=ftl-test, ftl-integration-test

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/442381
Original-Revision: bbb229e056563c099ee0d1ac1f73cac604ee4aab
GitOrigin-RevId: 8ef3d5392fecf3a882776c8ed3be7b786cd395be
Change-Id: I2c8b6a5f41c111c444bd4fb851e7c6a1947fe893