[roll] Roll fuchsia [bt][rfcomm] Handle DisconnectedMode responses
Adds support to handle DM responses from the remote peer. A DM response
is typically received as a form of rejection or error. For a user DLC, a
DM response typically means the DLC is in a bad state and should be
closed. For the mux control DLC, a DM response is typically sent if
the remote peer rejects multiplexer startup.

There is also a special case in which the peer can respond to a
Parameter Negotiation Mux command with a DM or DISC  over the DLC
referred to in the PN command. This special case is defined in RFCOMM
5.5.3. Adds support for this special case + tests.

Original-Bug: 59585
Test: Added to bt-rfcomm-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/444215
Original-Revision: dc9f8b1389d42c7403bb1b00f9b93d807c872322
GitOrigin-RevId: 70a16bbbcfc1467462c2d61c617bd44470d36ab7
Change-Id: I505de51d7bc568c5a3470b85e4fbb507d4c25841