[roll] Roll fuchsia [timekeeper] Update inspect to work in root(covariance).
We were forced to work with square root of covariance in Cobalt but it
is generally better - more intuitive and expressable as a zx::Duration.

Since we already need to perform the square root for cobalt, this CL
moves inspect over to also reporting the square root.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/444612
Original-Revision: f6b90b438cf6fd3d8b180d6da990932fdfb9bc6c
GitOrigin-RevId: 42155bbd38b95bf942344c3391078e84fc00d9bc
Change-Id: If47c1d9149739ab844a4ee868e7b3d9af7fdcb3d