[roll] Roll fuchsia Revert "[swd-e2e-tests] Retry artifacts tool failures"
This reverts commit e7d6da647bc626938d86a83b59261455a86fa47f.

Reason for revert: Infra has updated the `artifacts` prebuilt to use
a version of the storage API that handles the timeouts properly. This
patch was a band-aid to paper over that behavior -- there are likely
other failure modes where retries are incorrect, so best to leave all
of that up to the tool rather the consumer. Fixed in fxb/62502.

Original change's description:
> [swd-e2e-tests] Retry artifacts tool failures
>
> The artifacts prebuilt uses an old google-cloud-go storage API that
> doesn't handle retries on transient read failures. This wraps that
> command in a fixed-count retry loop such that transient issues won't
> cause test failures.
>
> Fixed: 62038
> Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/442054

TBR=etryzelaar@google.com,dhobsd@google.com,zkbaum@google.com

# Not skipping CQ checks because original CL landed > 1 day ago.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/445225
Original-Revision: 32279289b2445dfaa426a5142b06bd9bafb329db
GitOrigin-RevId: 1c6a51d18ff3b0fc84a423caafb6eedb773c8975
Change-Id: I7e5a1b83b4d868fcb2e164bff6df47c1ab854b78