[roll] Roll fuchsia [audio_core][tests] Fix windowing function
The Tukey windowing function is used during the clock-synchronization
integration tests. The function is passed an AudioBufferSlice,
processes it and returns an AudioBuffer with the results. This CL
fixes how the processing uses buffer vector indexing so that the
correct AudioBuffer is returned.

Test: test-only, audio-clocks-pipeline-test

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/445641
Original-Revision: 61e83a800b812f5f2a1a5ec8b497029919e2f3fa
GitOrigin-RevId: 5d1a3ec0b813b8958d0ec6210e17123ba724b7af
Change-Id: I13766e8e3136b146c3bafd85055427d4bc664c14