[roll] Roll fuchsia [testsharder][testrunner] stop on failure for multiply shards
The main utility of multiply is to check whether a test reliably
passes. Once a test has failed once, we know it does not reliably pass.
We can save bot and developer time by exiting early.

This new behavior is not very important and maybe even counter to
expectations for manual uses of multiply shards, but it will become
very important one we automatically multiply affected tests. The
affected tests are the ones most likely to deterministically fail, and
it's a big waste of time to run a test many times when it fails
deterministically.

If we get requests for the old "Keep Going" behavior, we can design a
way to bring that back for manual uses.

Original-Bug: 63143
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/444926
Original-Revision: 0a9387b844d32c175d257150674d723d765f8474
GitOrigin-RevId: 6a8bd498753b4b9e7d62f477e21be116893030a8
Change-Id: Ic991619b4c58f2f5c1eb332b1224ec1b943629db