[roll] Roll fuchsia [components] Add logging integration test
Verify that when there is a component which uses a LogSink
we see a capability route for it. Since the component never tries
to use the LogSink itself we assume this is because component
manager requested the capability on behalf of the component.

Original-Bug: 63000

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/445635
Original-Revision: 1c263dfc1f0b548c39d831c0eb2006483397b9e8
GitOrigin-RevId: 0846aeb34f0ea0af33156ad5d092e57179162f3b
Change-Id: If3fbdd8e41c1f839a75499d3aa5a417ed02c054f