[roll] Roll infra/recipes [flutter] Roll dart-pkg repo to correct revision
Previously, the flutter_with_dependents_roller always considered local
HEAD of the dart-pkg repo to point to the same revision hash as remote
HEAD after a roll attempt. However, this is a correct assumption *only*
if no changes were rolled. If some changes were committed and
successfully rolled, the resulting remote HEAD will generally have a
*different* revision from local HEAD because fuchsia Gerrit uses a
REBASE_ALWAYS merge strategy. This causes Gerrit to rebase CLs and add
some extra footers before submitting, which results in a different
commit hash.

Because of this, flutter-with-deps-roller was consistently failing every
time it rolled new dart-pkg changes, because it would then try to update
fuchsia's pinned version of that repo to point to a revision that only
existed locally, resulting in checkout failures in CQ.

The fix is to retrieve the revision from the final change-details
response from Gerrit after a dart-pkg roll has landed, and pin the repo
to *that* revision instead of local HEAD. If no roll dart-pkg roll is
necessary, we'll fall back to the old behavior of using local HEAD,
which should correspond to remote HEAD because we haven't made any local
changes.

Original-Bug: 62868
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/445795
Original-Revision: d352c83e129557ce102f78c75821d8ddfa6c82ae
GitOrigin-RevId: 465e5ae19f813dbad54f5b52ce5d47879d586334
Change-Id: Iaed9ebad041ba040453799ed6f8348a3a3ddb0a1