[roll] Roll fuchsia [zbitl] Introduce zbitl::Image
This change introduces zbitl::Image, which represents a "view" that also
permits operations that modify underlying ZBI (we conveniently look the
other way from View<>::iterator::EditHeader). In particular, we
introduce two versions of an operation to append new items: a partial,
more general you-can-write-the-payload-later version and a one-shot
one in which the provided header and data are written to the underlying
storage up front. (The latter is only efficient/sensible when the
desired payload is already in an simple, in-memory format.

Original-Bug: 57244
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/435260
Original-Revision: 5037d7006939c448305d6edb4d371aa2671bb1ae
GitOrigin-RevId: 35df2c772e51e89f5b65a4980f2a6eb60ff75538
Change-Id: I71ba7dd1396bed16b425b44bcc67dec7a7623740