[roll] Roll infra/recipes [cache] Only remove guard in try/else case.
Unconditionally removing the guard file in the finally case yields
errors because the guard may exist in a with block where CWD is
specified as a subdir of the cache. Even though file#remove() is
intended to succeed if the target file is already removed, operating
within a non-existant CWD is an error.

Original-Bug: 63129
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/446175
Original-Revision: 44f8d78840bb429dff56b8b30ac26472140b3aac
GitOrigin-RevId: 1af39792b04b3cd562d1b93f9f5bac0681256e49
Change-Id: I38e3a77de7796f9b31901b56d2811dc6c72b6d39