[roll] Roll fuchsia [wlan][policy] Clean up existing connection on new Connect()
When a new Connect() request comes in to the interface manager, we may
re-use an existing client interface. If this client interface had
an existing state machine, we should either:
- re-use the existing state machine and take no further action
- clean up the existing state machine

Prior to this change, the existing state machine was unconditionally
dropped without normal clean-up, which meant there would never be an
update to listeners to notify them that the connection maintained by
that state machine was being terminated. This was largely invisible
when the new connection was for the same network as the prior connection.

Many revisions ago, the deduping of identical Connect() request was
handled by the client state machine itself. With the refactor that made
client state machines particular to one single network, this deduping
functionality was lost and is now recreated in this CL.

Original-Bug: 63086
Test: fx test wlancfg-tests
Multiply: wlancfg-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/444955
Fuchsia-Auto-Submit: Marc Khouri <mnck@google.com>
Original-Revision: 359dfa80ad1d2e9dd93bab202a46af3865dfa4a5
GitOrigin-RevId: 183dfb0df43d64f751a75795474fbfc9d450af2b
Change-Id: I380a57c3ce4b29ecfa70326f84188802bbfb4706