[roll] Roll fuchsia [blobfs] Add oldest_revision to superblock
This implements the policy in //src/storage/docs/versioning.md for
tracking the oldest revision of the software that has touched the data.
The current revision is set to 1 so we can differentiate headers
produced before this change. The format is backwards-compatible with old
versions of the driver.

Similarly to was done for FVM in fxrev.dev/442564, rename the `version`
field to `format_version`.

Tested: New unit tests.
Original-Bug: 62593
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/444923
Original-Revision: 352d25283ef9e9674eace5de82c4920e730c77b8
GitOrigin-RevId: 4707be0b5b206f88c49639a31b968747ca661f5e
Change-Id: I5715cf2f051a64e700b86d85ebba9c2f2d6816dc