[roll] Roll fuchsia [weave] Refresh end points when the interface goes online/offline.
Whenever a device connects to an AP, the wifi interface acquires an
IP address and goes to online state. When the device disconnects from
an AP, the wifi interface loses IP and goes to offline state. In both
cases, refresh the weave endpoints so that stale endpoints are closed
and new endpoints are be created. This ensures the endpoints are updated
according to network state.

Original-Bug: 62872
Test: fx test weavestack-adaptation-unittests.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/443956
Original-Revision: a93c9e7aa63f3be18ba1cf165352f63039929c2c
GitOrigin-RevId: 64d03ef7de6c4f30068bf07d1b64cafedda1cb41
Change-Id: I6004a24369a126597d1b48a9a17069f763fc9d33