[roll] Roll fuchsia [check-licenses] Adding missing license patterns
This CL modifies some of our patterns, and adds many more
to reduce the number of files check-licenses fails to parse
in the fuchsia tree.

This also adds a list of "remaining_unlicensed_files" to main.go.
It's a large list of mostly manual work that we will burn down
as time permits.

We'd like to get the verification enabled, to ensure new unlicensed
files aren't added to the tree, and that list allows us to flip
the switch sooner.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/445643
Original-Revision: e87d050098a39c8c325a5f88c0a07913ea49b924
GitOrigin-RevId: d42a82b7ca14c2bab5d30950c7882673cb232e2e
Change-Id: I754a4b89a07700ec55b0a4d331834b231677a8c2