[roll] Roll fuchsia [timekeeper] Refactor clock correction.
This is a precursor to considering skewing the clock.

Behavior is close but not identical with this change. There are a few
corner cases that lead to a slightly different result, e.g. a correction
of less than 1ms previously would have skipped writing to the RTC but
now it doesn't.

There are also more explicit panics for the userspace clock failure cases
which previously led to a log error and skipping to the next iteration.
This seems reasonable given the deterministic clock failure modes and
the current importance of the userspace clock to timekeeper's role.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/445307
Original-Revision: eaf4066d54f6fc6e53dac959fe65b287123a4847
GitOrigin-RevId: d125ee466da4ce18331f61573477eecb73bd70e7
Change-Id: Iae72f20c8b7f4470d4956156b2359de4b5ec322d