[roll] Roll fuchsia [bootserver] Fix image paths to be relative to image manifest.
If using the images on the filesystem, bootserver would use the relative
paths of the images but make it relative to the current working
directory. This wasn't a problem because in our use cases, the image
manifest and images were always located in the directory that bootserver
was called from. However, this change will make it so bootserver can be
called from anywhere with the path to the image manifest and the image
paths defined in the manifest only need to be relative to the directory
of the manifest.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/445994
Original-Revision: 50cd21af5363a7635201d1eca5029cf885644861
GitOrigin-RevId: c48ad8ea5390fbc5c4808cedee5209a3de1ee212
Change-Id: I184b1221bee6017821237deaf6c2863fe14c1467