[roll] Roll infra/recipes [testing] Delete results_dir_on_host
There's nothing special about this directory anymore, so just create a
new temporary directory whenever we need to extract a task's test
results.

Also, have `extract_test_results` construct and return the test results
directory itself, instead of taking the directory as a parameter. The
function assumes that the test results dir does not yet exist, and its
easier to enforce that for a directory constructed within the function
than to validate it for a directory passed into the function.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/445695
Original-Revision: 66b1041f29ad8b62ed8aee57995e52c927aa6714
GitOrigin-RevId: fd2eb5602bc9ecb3feae562474ee36dcd4eca9e5
Change-Id: Ie9b49de1c03f24a673575e8cdd75732ce7e222af